fufox/requirements.txt
django
django-crispy-forms
django_star_ratings
IMDbPY
django-markdown2
django-comments-dab
django-widget-tweaks
django-allauth
beautifulsoup4
django-elasticsearch-dsl==6.4.2
google
